compose/auth-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
redis==5.0.7
PyJWT==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6
pydantic==2.5.0
